Add frontend function for new reviews
Also chane the switch in ConsoleApplication to actually call the wanted
function. Add use of new validator class in the callUp functions.
And need to add getCustomer and getReview to the interface because it
needs to be calles in the addNewReview callUp function to give it to the
validato^r